backend/requirements.txt
django
djangorestframework
pymongo
python-dotenv
django-cors-headers
